Add files via upload
This is just a little script to help launch applications from the command line and detaching their child process from the parent process. Place in $PATH for best results.

"x" is intended to be a prefixing script to use in your shell aliases. Example: "alias your-alias="x your-program-of-choice"

I recommend keeping this as-is. Dropdown/Scratchpad terminals are the best use case.