Not ready for use yet, just backing it up here
I just wanted to work on my own over-the-top configuration tree scheme for i3 in hopes to make full use of its wildcard include function. It is very "extra" and may not be your cup of tea, but this is how I want to achieve small file size for more complex configuration parameters that require frequent changes or more complex changes that may involve creating larger configuration files.